tgp_synth_pipeline/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tgp-synth"
version = "0.1.0"
description = "Therapeutic Generation Pipeline (TGP) - text->audio->video multimodal synthesis"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "TGP Contributors"}]
license = {text = "Apache-2.0"}

dependencies = [
  "typer>=0.12.3",
  "rich>=13.7.1",
  "pydantic>=2.6.0",
  "PyYAML>=6.0.1",
  "tqdm>=4.66.0",
  "numpy>=1.26.0",
]

[project.optional-dependencies]
rag = [
  "torch>=2.2.0",
  "torchvision>=0.17.0",
  "open_clip_torch>=2.24.0",
  "opencv-python>=4.8.0",
  "pillow>=10.0.0",
  "faiss-cpu>=1.8.0",
]

[project.scripts]
tgp-synth = "tgp_synth.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
